requirements.txt
discord
boto3
requests
python-dotenv
table2ascii
yt-dlp